requirements.txt
Django==4.1.7
psycopg2-binary==2.9.5
crispy-bootstrap5==0.7
django-tables2==2.5.3
Pillow==9.4.0
djangorestframework==3.14.0

pytest>=7.2.2
pytest-django>=4.5.2
scrapy==2.8.0

products==1.1.2
requests==2.28.2
Faker==18.4.0

factory-boy>=3.2.1
pytest-factoryboy>=2.5.0
gunicorn>=20.1.0
flake8>=6.0.0
isort>=5.12.0
black>=23.3.0